+++ src/pss.rs
@@ -36,8 +36,8 @@ use crate::algorithms::pad::{uint_to_be_pad, uint_to_zeroizing_be_pad};
 use crate::algorithms::pss::*;
 use crate::algorithms::rsa::{rsa_decrypt_and_check, rsa_encrypt};
 use crate::errors::{Error, Result};
-use crate::keytraits::PublicKeyParts;
-use crate::padding::SignatureScheme;
+use crate::traits::SignatureScheme;
+use crate::traits::PublicKeyParts;
 use crate::{RsaPrivateKey, RsaPublicKey};
 
 #[cfg(feature = "getrandom")]
