+++ tests/pkcs1.rs
@@ -3,7 +3,8 @@
 use hex_literal::hex;
 use rsa::{
     pkcs1::{DecodeRsaPrivateKey, DecodeRsaPublicKey, EncodeRsaPrivateKey, EncodeRsaPublicKey},
-    PrivateKeyParts, PublicKeyParts, RsaPrivateKey, RsaPublicKey,
+    traits::{PrivateKeyParts, PublicKeyParts},
+    RsaPrivateKey, RsaPublicKey,
 };
 
 #[cfg(feature = "pem")]
